Add uniqueness check to --package input
CLOUDBLD-11120

If the user specifies the same --package twice, process it only once. If
the user specifies conflicting packages (currently impossible), return
an error.
